Group1_proj/requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
langchain-chroma>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchainhub
chromadb>=0.4.24
pandas
python-dotenv
fastapi
uvicorn
streamlit
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai
sentence-transformers